[#20] Add scroll functionality to response textboxes
When a response length exceeds a certain length, the text box does not scroll to bottom to show the remaining text.

There are two ways to fix this:
1. Display the scroll bar and let the user scroll to the bottom. When it reaches the bottom, the scroll bar will automatically scroll to the bottom.
2. Automatically scroll to the bottom when the response text is generated.

This PR implements the second option since it's cumbersome to scroll to the bottom manually.

When the response text is generated, `element.scrollTop = element.scrollHeight` is triggered to scroll to the bottom of the text box.
It's not the most efficient way to do this, but it's the simplest way to do it.