[#37] Store ELO ratings in DB after calculation
This change adds logic to store ELO ratings in the database after they are calculated.

Previously, we calculated and loaded the ratings without storing them. Now, we store them for future use.

This change doesn't affect current operations as we are not using the stored data yet. However, it sets the groundwork for future optimizations where we will use the stored ratings to avoid recalculating ELO scores for already calculated battles.